codes/anti-barrage/requirements.txt
tensorflow==2.2.0
pixellib==0.7.1
scipy==1.4.1
scikit-image==0.19.0
numpy==1.18.0
matplotlib==3.5.0
protobuf==3.20.0
pillow==9.0